Extract 'points_to_center_of_mass' into 'core' package
